Connecting headers into one
