--- frontend/src/components/header/header.module.scss
@@ -1,18 +1,0 @@
-.header {
-  position: relative;
-  z-index: 2;
-  background: var(--white);
-}
-
-.headerInner {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: var(--header-height);
-}
-
-.container {
-  display: flex;
-  align-items: center;
-}
